Deep changes in how the different objects are stored in eachother, i think it makes more sense now.
Not working yet, though.

The algorithm to check for winning move has been moved to Game in an
ugly way. This needs new methods in GameBoard, some of which are made
(getSquare(...), etc), but these imply new changes that have yet to be
made. But that will have to wait for another day.