Still on the right direction, not quite there yet.
Today I Learned you cannot initialize an array with an enhanced for
loop. Which I'd rather have learnt some other way, but I guess it's OK.

I've also found a good reason why it's better to prevent Exceptions than
to catch them. I've come to see the reason (one of the reasons, at the
very least) why my win detection algorithm isn't working is because
every time there's an issue with the current row or col, an exception 
is thrown and the instruction continue; is executed. And everything goes
away.

Now that I understand the problem it shouldn't take too long to fix, but 
that'll have to wait for another day.